Refonte de la phase de dialogue
- DialogueSO actuellement inutilisable.
- Retrait des conditions d'accès aux approches.
- Ajout du test de la tentative d'approche.
