+++ Gourou Manager/Assets/Scripts/Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 96c6c49adb2439544915352106864fbf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
