+++ Gourou Manager/Assets/Scripts/Dialogue/ApproachSO.cs.meta
