+++ Gourou Manager/Assets/Scripts/Dialogue/DialogueSO.cs.meta
